Finished adding tests and docstrings

+++ src/main/java/Transaction.java
@@ -18,6 +18,6 @@ public class Transaction {
              total += price;
          }
 
-         return Math.round(total * 100) / 100;
+         return ((float)Math.round(total * 100) / 100);
     }
 }
